requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
asgiref==3.4.1; python_version >= '3.6'
beautifulsoup4==4.10.0; python_version >= '3.1'
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
django-appconf==1.0.5; python_version >= '3.6'
django-autocomplete-light==3.8.2
django-compressor==2.4.1
django-crispy-forms==1.13.0
django-filter==21.1
django-forms-bootstrap==3.1.0
django-imagekit==4.0.2
django-model-utils==4.2.0
django-ordered-model==3.4.3
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-sass-processor==1.1
django==3.2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3; python_version >= '3'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.0.2; python_version >= '3.6'
libsass==0.21.0
markupsafe==2.0.1; python_version >= '3.6'
packaging==21.0; python_version >= '3.6'
pilkit==2.0
pillow==8.4.0
psycopg2==2.9.1
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
pytz==2021.3
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.1.0
ruamel.yaml.clib==0.2.6; platform_python_implementation == 'CPython' and python_version < '3.10'
ruamel.yaml==0.17.16; python_version >= '3'
six==1.16.0
soupsieve==2.2.1; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
